Now able to change the number of items on a page #130.
Current implementation does not make any changes to the View when the
number of items per page is changed; not even to update the number of
pages. Instead, the pagaination footer is updated with the rest of it
when the Next button is pressed.

This was done to prevent the Next button from disappearing if the user
increased the page size such that it is on the last page, but not all
the items that should be on its last page have been rendered yet.

For example:
datumsId has 5 ids.
perPage starts out at 3. Since there are 3 DatumLatexViews, it calculates
it's on page 1 of 2; therefore, the Next button is visible.
Increase perPage to 9. Since there are 3 DatumLatexViews, it calculates
it's on page 1 of 1; therefore, the Next button is invisible and there
is no way to display the last two datumIds.