creating a backbone model and putting twitter data
into it. can use it to look up existing users. for #204
TODO try to add permissions db in couch, and test the code to see if twitter  usernames now show up in docs in the users db
TODO figure out how to tie usernames to the accounts created in users db, why a view for twitter users? why not a view for all users? investigate everyauth conventions more