Made modals hidden by default. Used the correct modal("show") method to display the modal in the router rather than the *very* incorrect show() method.
The show() method will just display the <div> but it will not act like
a modal anymore (i.e., it won't close if you push the close button).

No longer need to explicitly hide the modals in the hideEverything()
method because they disappear as soon as you click elsewhere on the screen.