Working on differentiatng perminaten and temproary data lists created by the leftisdeeditdata list. in reality i think we need to do the same thing that import does, that is that import has its own dashboard with its own data list. search should have a dashboard with a data list, and if the user clicks save on that data list, then it saves the temp data list into a permanent one and sets taht as the curent one and brings the user back to their normal dashboard.
but this wont work because then we cant load stuff into the right side data list if its in an active serach mode while we are entering data. right? or woudl it actually be easier because there will be a dedicated data list that is the result of a curren search.